fix redundant and failed bytes overflow check

+++ src/torrent.cpp
@@ -11108,10 +11108,10 @@ bool is_downloading_state(int const st)
 		TORRENT_ASSERT(static_cast<int>(reason) >= 0);
 		TORRENT_ASSERT(static_cast<int>(reason) < static_cast<int>(waste_reason::max));
 
-		if (m_total_redundant_bytes <= std::numeric_limits<std::int32_t>::max() - b)
+		if (m_total_redundant_bytes <= std::numeric_limits<std::int64_t>::max() - b)
 			m_total_redundant_bytes += b;
 		else
-			m_total_redundant_bytes = std::numeric_limits<std::int32_t>::max();
+			m_total_redundant_bytes = std::numeric_limits<std::int64_t>::max();
 
 		// the stats counters are 64 bits, so we don't check for overflow there
 		m_stats_counters.inc_stats_counter(counters::recv_redundant_bytes, b);
@@ -11122,10 +11122,10 @@ bool is_downloading_state(int const st)
 	{
 		TORRENT_ASSERT(is_single_thread());
 		TORRENT_ASSERT(b > 0);
-		if (m_total_failed_bytes <= std::numeric_limits<std::int32_t>::max() - b)
+		if (m_total_failed_bytes <= std::numeric_limits<std::int64_t>::max() - b)
 			m_total_failed_bytes += b;
 		else
-			m_total_failed_bytes = std::numeric_limits<std::int32_t>::max();
+			m_total_failed_bytes = std::numeric_limits<std::int64_t>::max();
 
 		// the stats counters are 64 bits, so we don't check for overflow there
 		m_stats_counters.inc_stats_counter(counters::recv_failed_bytes, b);
